+++ e2e/e2etests/e2etests.go
@@ -100,12 +100,13 @@ const (
 	 Admin tests
 	 Test admin functionalities
 	*/
-	TestDepositEtherLiquidityCapName = "deposit_eth_liquidity_cap"
-	TestMigrateChainSupportName      = "migrate_chain_support"
-	TestPauseZRC20Name               = "pause_zrc20"
-	TestUpdateBytecodeZRC20Name      = "update_bytecode_zrc20"
-	TestUpdateBytecodeConnectorName  = "update_bytecode_connector"
-	TestRateLimiterName              = "rate_limiter"
+	TestDepositEtherLiquidityCapName  = "deposit_eth_liquidity_cap"
+	TestMigrateChainSupportName       = "migrate_chain_support"
+	TestPauseZRC20Name                = "pause_zrc20"
+	TestUpdateBytecodeZRC20Name       = "update_bytecode_zrc20"
+	TestUpdateBytecodeConnectorName   = "update_bytecode_connector"
+	TestRateLimiterName               = "rate_limiter"
+	TestCriticalAdminTransactionsName = "critical_admin_transactions"
 
 	TestMigrateTSSName = "migrate_TSS"
 
@@ -557,6 +558,12 @@ var AllE2ETests = []runner.E2ETest{
 		"test sending cctxs with rate limiter enabled and show logs when processing cctxs",
 		[]runner.ArgDefinition{},
 		TestRateLimiter,
+	),
+	runner.NewE2ETest(
+		TestCriticalAdminTransactionsName,
+		"test critical admin transactions",
+		[]runner.ArgDefinition{},
+		TestCriticalAdminTransactions,
 	),
 	/*
 	 Special tests