+++ zetaclient/zetacore/client_vote.go
@@ -46,9 +46,7 @@ func (c *Client) PostVoteBlockHeader(
 func (c *Client) PostVoteGasPrice(
 	ctx context.Context,
 	chain chains.Chain,
-	gasPrice uint64,
-	supply string,
-	blockNum uint64,
+	gasPrice uint64, priorityFee, blockNum uint64,
 ) (string, error) {
 	// get gas price multiplier for the chain
 	multiplier := GasPriceMultiplier(chain)
@@ -56,7 +54,7 @@ func (c *Client) PostVoteGasPrice(
 	// #nosec G115 always in range
 	gasPrice = uint64(float64(gasPrice) * multiplier)
 	signerAddress := c.keys.GetOperatorAddress().String()
-	msg := types.NewMsgVoteGasPrice(signerAddress, chain.ChainId, gasPrice, supply, blockNum)
+	msg := types.NewMsgVoteGasPrice(signerAddress, chain.ChainId, gasPrice, priorityFee, blockNum)
 
 	authzMsg, authzSigner, err := WrapMessageWithAuthz(msg)
 	if err != nil {