+++ changelog.md
@@ -34,6 +34,7 @@
 * [2465](https://github.com/zeta-chain/node/pull/2465) - add Solana inbound SOL token observation
 * [2497](https://github.com/zeta-chain/node/pull/2416) - support for runtime chain (de)provisioning
 * [2518](https://github.com/zeta-chain/node/pull/2518) - add support for Solana address in zetacore
+* [2483](https://github.com/zeta-chain/node/pull/2483) - add priorityFee (gasTipCap) gas to the state
 
 ### Refactor
 
@@ -63,6 +64,7 @@
 * [2395](https://github.com/zeta-chain/node/pull/2395) - converge AppContext with ZetaCoreContext in zetaclient
 * [2428](https://github.com/zeta-chain/node/pull/2428) - propagate context across codebase & refactor zetacore client
 * [2464](https://github.com/zeta-chain/node/pull/2464) - move common voting logic to voting.go and add new function VoteOnBallot
+* [2515](https://github.com/zeta-chain/node/pull/2515) - replace chainName by chainID for ChainNonces indexing
 
 ### Tests
 
@@ -84,7 +86,7 @@
 * [2549](https://github.com/zeta-chain/node/pull/2459) - add separate accounts for each policy in e2e tests
 * [2415](https://github.com/zeta-chain/node/pull/2415) - add e2e test for upgrade and test admin functionalities
 * [2440](https://github.com/zeta-chain/node/pull/2440) - Add e2e test for TSS migration
-
+* [2473](https://github.com/zeta-chain/node/pull/2473) - add e2e tests for most used admin transactions
 
 ### Fixes